feat(timeService): 시간 삭제 API에서 해당 시간에 예약이 있을 경우 예외 처리

+++ src/main/java/roomescape/service/TimeService.java
@@ -3,15 +3,18 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.controller.dto.ReservationTimeAddRequest;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.TimeDao;
 import roomescape.domain.ReservationTime;
 
 @Service
 public class TimeService {
     private final TimeDao timeDao;
+    private final ReservationDao reservationDao;
 
-    public TimeService(TimeDao timeDao) {
+    public TimeService(TimeDao timeDao, ReservationDao reservationDao) {
         this.timeDao = timeDao;
+        this.reservationDao = reservationDao;
     }
 
     public ReservationTime addReservationTime(ReservationTimeAddRequest request) {
@@ -30,6 +33,9 @@ public class TimeService {
     }
 
     public void removeReservationTime(Long id) {
+        if (reservationDao.isExistByTimeId(id)) {
+            throw new IllegalArgumentException("해당 시간에 예약이 존재합니다.");
+        }
         timeDao.delete(id);
     }
